Refactor cart and dashboard cubits to calculate and display total price of items in cart
The changes in this commit refactor the CartCubit and DashCubit to calculate the total price of items in the cart and display it in the TotalPriceDetailsWidget. The sumOfItemsPrice method has been moved from DashCubit to CartCubit and modified to take in the price of the item being added to the cart. The TotalPriceDetailsWidget has been updated to display the total price calculated in