Decrease item price in cart when decreasing quantity
This commit fixes a bug in the `CartCubit` where the item price was being increased instead of decreased when the quantity of an item was decreased. The bug was fixed by subtracting the item price from the total price instead of adding it. The `sumOfItemsPriceByDecreaseIcon` method was also removed as it was not being used.