Add ShippingToItem class and functionality to add shipping addresses to a list
This commit adds a new class called ShippingToItem which contains the address and phone number of a shipping address. It also adds functionality to the ShippingToCubit to add a new ShippingToItem to a list when the user saves a new address. The list of ShippingToItems is then displayed in the ShipToScreen using a ListView. The ShiptolistItemWidget is updated to display the address and phone number from the ShippingToItem instead of using the addressController and phoneController.