Add branch selection dropdown to register screen
This commit adds a dropdown menu to the register screen that allows the user to select their branch. The selected branch is saved to local storage using the CacheHelper class. The dropdown menu is populated with two options: "Beniseuf" and "Cairo". The commit also includes some minor code cleanup and refactoring.